# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: OpenAI AI chat operations
  - name: sms
    description: SMS sending via Twilio
  - name: payments
    description: Payment processing via Stripe
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

  /sms/send:
    post:
      operationId: sendSms
      tags: [sms]
      summary: Send an SMS message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsInput"
      responses:
        "200":
          description: SMS sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /payments/create-intent:
    post:
      operationId: createPaymentIntent
      tags: [payments]
      summary: Create a Stripe payment intent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentIntentInput"
      responses:
        "200":
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntentResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /payments/webhook:
    post:
      operationId: stripeWebhook
      tags: [payments]
      summary: Handle Stripe webhook events
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResult"
  /payments/history:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payment history
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRecord"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    SmsInput:
      type: object
      properties:
        to:
          type: string
          description: Recipient phone number in E.164 format (e.g. +15551234567)
        message:
          type: string
          description: The SMS message body
      required:
        - to
        - message
    SmsResult:
      type: object
      properties:
        success:
          type: boolean
        sid:
          type: string
        status:
          type: string
      required:
        - success
        - sid
        - status

    PaymentIntentInput:
      type: object
      properties:
        amount:
          type: integer
          description: Amount in cents (e.g. 999 = $9.99)
        currency:
          type: string
          default: usd
        description:
          type: string
      required:
        - amount
    PaymentIntentResult:
      type: object
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
        amount:
          type: integer
        currency:
          type: string
      required:
        - clientSecret
        - paymentIntentId
        - amount
        - currency
    PaymentRecord:
      type: object
      properties:
        id:
          type: integer
        stripePaymentIntentId:
          type: string
        amount:
          type: integer
        currency:
          type: string
        status:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - stripePaymentIntentId
        - amount
        - currency
        - status
        - createdAt
    WebhookResult:
      type: object
      properties:
        received:
          type: boolean
      required:
        - received
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error