# lib/api-zod/src/generated/types/paymentIntentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentIntentInput {
  /** Amount in cents (e.g. 999 = $9.99) */
  amount: number;
  currency?: string;
  description?: string;
}