# lib/api-zod/src/generated/types/smsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SmsInput {
  /** Recipient phone number in E.164 format (e.g. +15551234567) */
  to: string;
  /** The SMS message body */
  message: string;
}